moved partition, modified tests, documents
