+++ modules/core/doc/core.rst
@@ -12,7 +12,6 @@ core. The Core Functionality
     operations_on_arrays
     xml_yaml_persistence
     old_xml_yaml_persistence
-    clustering
     utility_and_system_functions_and_macros
     opengl_interop
     ipp_async_converters
