+++ modules/java/android_test/src/org/opencv/test/core/CoreTest.java
@@ -792,64 +792,6 @@ public class CoreTest extends OpenCVTestCase {
         assertEquals(0.3819660544395447, cond, EPS);
     }
 
-    public void testKmeansMatIntMatTermCriteriaIntInt() {
-        Mat data = new Mat(4, 5, CvType.CV_32FC1) {
-            {
-                put(0, 0, 1, 2, 3, 4, 5);
-                put(1, 0, 2, 3, 4, 5, 6);
-                put(2, 0, 5, 4, 3, 2, 1);
-                put(3, 0, 6, 5, 4, 3, 2);
-            }
-        };
-        TermCriteria criteria = new TermCriteria(TermCriteria.EPS, 0, EPS);
-        Mat labels = new Mat();
-
-        Core.kmeans(data, 2, labels, criteria, 1, Core.KMEANS_PP_CENTERS);
-
-        int[] first_center = new int[1];
-        labels.get(0, 0, first_center);
-        final int c1 = first_center[0];
-        Mat expected_labels = new Mat(4, 1, CvType.CV_32S) {
-            {
-                put(0, 0, c1, c1, 1 - c1, 1 - c1);
-            }
-        };
-        assertMatEqual(expected_labels, labels);
-    }
-
-    public void testKmeansMatIntMatTermCriteriaIntIntMat() {
-        Mat data = new Mat(4, 5, CvType.CV_32FC1) {
-            {
-                put(0, 0, 1, 2, 3, 4, 5);
-                put(1, 0, 2, 3, 4, 5, 6);
-                put(2, 0, 5, 4, 3, 2, 1);
-                put(3, 0, 6, 5, 4, 3, 2);
-            }
-        };
-        TermCriteria criteria = new TermCriteria(TermCriteria.EPS, 0, EPS);
-        Mat labels = new Mat();
-        Mat centers = new Mat();
-
-        Core.kmeans(data, 2, labels, criteria, 6, Core.KMEANS_RANDOM_CENTERS, centers);
-
-        int[] first_center = new int[1];
-        labels.get(0, 0, first_center);
-        final int c1 = first_center[0];
-        Mat expected_labels = new Mat(4, 1, CvType.CV_32S) {
-            {
-                put(0, 0, c1, c1, 1 - c1, 1 - c1);
-            }
-        };
-        Mat expected_centers = new Mat(2, 5, CvType.CV_32FC1) {
-            {
-                put(c1, 0, 1.5, 2.5, 3.5, 4.5, 5.5);
-                put(1 - c1, 0, 5.5, 4.5, 3.5, 2.5, 1.5);
-            }
-        };
-        assertMatEqual(expected_labels, labels);
-        assertMatEqual(expected_centers, centers, EPS);
-    }
-
     public void testLineMatPointPointScalar() {
         int nPoints = Math.min(gray0.cols(), gray0.rows());
         Point point1 = new Point(0, 0);
